VCRテストの環境依存対策
