+++ tests/cassettes/test_json_to_pptx_analyze_tool_vcr.yaml
@@ -1,6 +1,6 @@
 interactions:
 - request:
-    body: '{"pptx_template_id": "19823658-6af5-4560-ab87-632da39511de"}'
+    body: '{"pptx_template_id": "TEMPLATE_ID"}'
     headers:
       Accept:
       - '*/*'
@@ -31,15 +31,24 @@ interactions:
         title"}]},{"position":2,"type":"Common","description":"Please use this slide
         to create content.","placeholders":[{"name":"title","description":"Input slide
         title here"},{"name":"content","description":"Input slide content here"},{"name":"image","description":"Input
-        slide image url here"}]}],"json_to_pptx_presentation_schema":{"$defs":{"Placeholder":{"properties":{"name":{"description":"The
-        key of the placeholder","title":"Name","type":"string"},"content":{"description":"The
-        content of the placeholder","title":"Content","type":"string"}},"required":["name","content"],"title":"Placeholder","type":"object"},"Slide":{"properties":{"type":{"description":"The
-        type of the slide","title":"Type","type":"string"},"placeholders":{"description":"The
-        placeholders of the slide","items":{"$ref":"#/$defs/Placeholder"},"title":"Placeholders","type":"array"},"speaker_notes":{"anyOf":[{"type":"string"},{"type":"null"}],"default":null,"description":"Additional
-        notes for the slide intended for the presenter. These notes provide extra
-        context or cues during the presentation and are not visible to the audience.","title":"Speaker
-        Notes"}},"required":["type","placeholders"],"title":"Slide","type":"object"}},"properties":{"slides":{"description":"The
-        slides of the presentation","items":{"$ref":"#/$defs/Slide"},"title":"Slides","type":"array"}},"required":["slides"],"title":"Presentation","type":"object"}}'
+        slide image url here"}]}],"json_to_pptx_presentation_schema":"{\"$defs\":
+        {\"Placeholder\": {\"properties\": {\"name\": {\"description\": \"The key
+        of the placeholder\", \"title\": \"Name\", \"type\": \"string\"}, \"content\":
+        {\"description\": \"The content of the placeholder\", \"title\": \"Content\",
+        \"type\": \"string\"}}, \"required\": [\"name\", \"content\"], \"title\":
+        \"Placeholder\", \"type\": \"object\"}, \"Slide\": {\"properties\": {\"type\":
+        {\"description\": \"The type of the slide\", \"title\": \"Type\", \"type\":
+        \"string\"}, \"placeholders\": {\"description\": \"The placeholders of the
+        slide\", \"items\": {\"$ref\": \"#/$defs/Placeholder\"}, \"title\": \"Placeholders\",
+        \"type\": \"array\"}, \"speaker_notes\": {\"anyOf\": [{\"type\": \"string\"},
+        {\"type\": \"null\"}], \"default\": null, \"description\": \"Additional notes
+        for the slide intended for the presenter. These notes provide extra context
+        or cues during the presentation and are not visible to the audience.\", \"title\":
+        \"Speaker Notes\"}}, \"required\": [\"type\", \"placeholders\"], \"title\":
+        \"Slide\", \"type\": \"object\"}}, \"properties\": {\"slides\": {\"description\":
+        \"The slides of the presentation\", \"items\": {\"$ref\": \"#/$defs/Slide\"},
+        \"title\": \"Slides\", \"type\": \"array\"}}, \"required\": [\"slides\"],
+        \"title\": \"Presentation\", \"type\": \"object\"}"}'
     headers:
       content-encoding:
       - gzip
